File Path Cleanup
