+++ Business_closure_rate_T_Test/rate_2019_2021.py
